requirements.txt
pywin32>=303
pandas>=1.4.2
Pillow>=8.4.0
pytesseract>=0.3.10
opencv-python>=4.7.0.72
openpyxl>=3.1.2
numpy>=1.22.3